factoryBurn: adnl: fix null address if ctrl-c [1/1]
PD#SWPL-160733

Problem:
has null address access if ctrl c in adnl mode
usb_burn erase flash failed
usb_burn no display UI

Solution:
check null pointer before access
fix usb_burn erase failed
add config for usb_burn UI

Verify:
s7_bh201

+++ bl33/v2023/drivers/usb/gadget/v3_burning/v3_common/v3_tool_buff_manager.c
@@ -88,7 +88,7 @@ int v3tool_buffman_img_verify_sha1sum(unsigned char *vrySum)
 	} else {
 		switch (imgFmt) {
 		case V3TOOL_PART_IMG_FMT_RAW: {
-			int thisVryLen = 0;
+			int64_t thisVryLen = 0;
 			static unsigned long _lastReportTick;
 
 			if (!vryLen)
